reorganizing classes for redistribution
