+++ tests/ProductTestModel.php
@@ -5,11 +5,11 @@ namespace NeptuneSoftware\Invoicable;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Str;
-use NeptuneSoftware\Invoicable\IsInvoicable\IsInvoicableTrait;
+use NeptuneSoftware\Invoicable\Traits\InvoicableTrait;
 
 class ProductTestModel extends Model
 {
-    use IsInvoicableTrait;
+    use InvoicableTrait;
 
     public $incrementing = false;
